chore: rename imageTtemplates to imageTemplates
