+++ apps/client/composables/main/shareImage/imageTemplates/tpl_1.ts
