+++ apps/client/composables/main/shareImage/imageTemplates/tpl_2.ts
